refactor: updating event filenames
