+++ src/app/(private)/(routes)/admin/events/components/event-client.tsx
